Critical Commit
- Authentication (login/register/logout) functional
- File upload functional
- File encryption via encryptionString POST
- File decryption on the fly
- Secure HTTP-only cookie implementation to securely store Bearer and userID credentials.

At this time the frontend is being served from a different domain for test purposes, simply change the links to your own you should be good to go.

Cookies will be served on the backend domain, NOT the frontend so adjust accordingly on your frontend.

If a user fetches an encrypted file(set true in DB) without a encryptionString present in the request, the server will return failure notifying that the file is encrypted.

Overall project IS functional from a backend perspective but some ideas such as:
- getAll() type of endpoint that fetches all currently uploaded file names and access links to create a type of searchable directory on the frontend.
- Similarly, the option to mark a file as private(so it will not be present in this directory)
- Allow a user to obtain a list of their individual uploaded file with access links

QOL Plans:
Remove unnecessary logging and move static addresses/credentials to the project properties for easy modification.
Review sloppy/redundant code and shrink where needed.